requirements.txt
azure-ai-documentintelligence>=1.0.0b3
azure-core>=1.30.0
openai>=1.0.0                     
pandas>=2.0
openpyxl>=3.0
python-dotenv>=1.0
azure-storage-blob>=12.18.0
azure-identity>=1.18.0
pypdf>=4.0.0